+++ mlflow/__init__.py
@@ -168,7 +168,7 @@ from mlflow.models import evaluate
 from mlflow.client import MlflowClient
 from mlflow.exceptions import MlflowException
 from mlflow.projects import run
-from mlflow.doctor import doctor
+from mlflow._doctor import doctor
 
 __all__ = [
     "ActiveRun",
